fix(MOR-2036): move the data-* vocabulary out of the workspace purity closure
presentation/languages/contract.ts is a member of the MOR-1077/1078/1079
workspace-purity closures, which reject any closure member naming
semantic/ in an import — type-only included, since the closure walker
reads source text and does not erase `import type`. The two semantic/
imports this ticket added tripped all three purity suites and left
`Tests (quick)` red at the previous head.

Fix: relocate the whole MOR-2036 vocabulary section to a new sibling
module, presentation/languages/state-vocabulary.ts, which is not
reachable from workspace/contract.ts's closure and may therefore read
semantic/ types the same one-way way projection.ts already does.
contract.ts itself is restored to zero imports (its pre-ticket shape)
plus a doc-comment pointer to the new file. Inverting the dependency
(defining the unions in contract.ts and having semantic/rx-tx-surface.ts
import them back) was considered and rejected: that reverses the
accepted presentation-depends-on-semantic direction (v3 ADR; April ADR's
"each layer depends only on the layer below, never upward").

Also corrects three false doc-comment claims flagged in review:
- "enforced here by inspection rather than by an eslint rule" — the
  three purity closures are the real, named enforcement.
- "the contract module cannot value-import from semantic/" — nothing
  forbids it; the value arrays are literals by choice, not prohibition.
- "presentation/ reads semantic/ types only, never runtime values" and
  the "same reason radio-view-model.ts re-declares MeterRfState"
  cross-reference — both false; six presentation/ test files do
  value-import semantic/, and MeterRfState's redeclaration is an
  unrelated intra-semantic/ cycle-avoidance.

And documents a review-found gap: TxPanel.svelte's data-fault was
miscategorized as boolean (it renders a fault CODE string, same camp as
AppGlobalHost.svelte) and its third, undocumented data-rf collision
('on' | 'off' | 'unknown') is now noted alongside RF_STATES.
TxPanel.svelte itself is untouched — out of scope by owner ruling.

Verification: purity suite 3 files/33 tests pass; npx vitest run
src/presentation/ src/semantic/ 94 files/2441 tests pass; full npx
vitest run 369 files/8081 tests pass; npm run check 0 errors/0
warnings (4601 files); npm run lint and npm run lint:boundaries clean.
